refactor(settings): extract standalone components to _components/
PR-1 of the settings-page split. settings/page.tsx is 2100+ lines
in a single component; the split pulls cohesive pieces out so the
file becomes reviewable and per-section state stops re-rendering
the whole tree.

This first slice moves the three pure-presentational helpers — no
settings-state coupling — into dashboard/src/app/settings/_components/:

- ToggleRow      — checkbox + label + help row
- PermColumn     — one column of the approval-policy grid
- ConfigFileCard — config-path card (owns only its copy-feedback state)

Pure relocation: identical JSX, no behavior change. Establishes the
_components/ directory the section extractions (Telemetry, Mobile,
Safety, Approval, AI Drivers, Bridge) will land in next.
