+++ dashboard/src/app/settings/page.tsx
@@ -11,6 +11,9 @@ import {
   subscribeToPush,
   unsubscribeFromPush,
 } from "@/lib/pushSubscription";
+import { ConfigFileCard } from "./_components/ConfigFileCard";
+import { PermColumn } from "./_components/PermColumn";
+import { ToggleRow } from "./_components/ToggleRow";
 
 interface StatusResponse {
   uptimeMs?: number;
@@ -1964,181 +1967,5 @@ const SettingsNavList = memo(function SettingsNavList({
   );
 });
 
-function ToggleRow({
-  id,
-  label,
-  help,
-  checked,
-  onChange,
-  disabled,
-  disabledReason,
-}: {
-  id: string;
-  label: string;
-  help: string;
-  checked: boolean;
-  onChange: (v: boolean) => void;
-  /**
-   * When true, the checkbox cannot be toggled by the user. Used by
-   * the kill-switch row when env-locked (#422 v2-I7).
-   */
-  disabled?: boolean;
-  /** Tooltip text shown on hover when disabled is true. */
-  disabledReason?: string;
-}) {
-  return (
-    <div
-      style={{ display: "flex", gap: 12, alignItems: "flex-start" }}
-      title={disabled ? disabledReason : undefined}
-    >
-      <input
-        id={id}
-        type="checkbox"
-        checked={checked}
-        disabled={disabled}
-        onChange={(e) => onChange(e.target.checked)}
-        style={{ marginTop: 2, cursor: disabled ? "not-allowed" : undefined }}
-      />
-      <label
-        htmlFor={id}
-        style={{ flex: 1, cursor: disabled ? "not-allowed" : "pointer" }}
-      >
-        <div
-          style={{
-            fontSize: "var(--fs-m)",
-            fontWeight: 500,
-            color: disabled ? "var(--fg-2)" : "var(--fg-0)",
-          }}
-        >
-          {label}
-        </div>
-        <div
-          style={{
-            fontSize: "var(--fs-s)",
-            color: "var(--fg-2)",
-            marginTop: 2,
-            lineHeight: 1.5,
-          }}
-        >
-          {help}
-          {disabled && disabledReason ? (
-            <div
-              style={{
-                marginTop: 4,
-                fontSize: "var(--fs-xs)",
-                color: "var(--fg-3)",
-                fontStyle: "italic",
-              }}
-            >
-              ⓘ {disabledReason}
-            </div>
-          ) : null}
-        </div>
-      </label>
-    </div>
-  );
-}
-
-function PermColumn({
-  tone,
-  title,
-  rules,
-}: {
-  tone: "ok" | "warn" | "err";
-  title: string;
-  rules: string[];
-}) {
-  return (
-    <div
-      style={{
-        background: "var(--bg-2)",
-        border: "1px solid var(--border-default)",
-        borderRadius: "var(--r-2)",
-        padding: 10,
-        minHeight: 80,
-      }}
-    >
-      <div style={{ display: "flex", alignItems: "center", gap: 6, marginBottom: 8 }}>
-        <StatusPill tone={tone}>{title}</StatusPill>
-        <span style={{ fontSize: "var(--fs-xs)", color: "var(--fg-3)" }}>{rules.length}</span>
-      </div>
-      {rules.length === 0 ? (
-        <div style={{ fontSize: "var(--fs-xs)", color: "var(--fg-3)" }}>—</div>
-      ) : (
-        <ul style={{ listStyle: "none", margin: 0, padding: 0, display: "flex", flexDirection: "column", gap: 3 }}>
-          {rules.slice(0, 8).map((r) => (
-            <li key={r} className="mono" style={{ fontSize: "var(--fs-xs)", color: "var(--fg-1)", wordBreak: "break-all" }}>
-              {r}
-            </li>
-          ))}
-          {rules.length > 8 && (
-            <li style={{ fontSize: "var(--fs-xs)", color: "var(--fg-3)" }}>+{rules.length - 8} more</li>
-          )}
-        </ul>
-      )}
-    </div>
-  );
-}
-
-function ConfigFileCard({ path }: { path: string }) {
-  const [state, setState] = useState<"idle" | "copied" | "blocked">("idle");
-
-  async function copy() {
-    try {
-      await navigator.clipboard.writeText(path);
-      setState("copied");
-      setTimeout(() => setState("idle"), 1500);
-    } catch {
-      // Clipboard API blocked (insecure context, headless, permission denied).
-      // Tell the user instead of silently no-op'ing.
-      setState("blocked");
-      setTimeout(() => setState("idle"), 2400);
-    }
-  }
-
-  const copied = state === "copied";
-  const blocked = state === "blocked";
-
-  return (
-    <div
-      style={{
-        marginTop: 16,
-        background: "var(--bg-2)",
-        border: "1px solid var(--border-default)",
-        borderRadius: "var(--r-2)",
-        padding: 10,
-      }}
-    >
-      <div style={{ fontSize: "var(--fs-2xs)", fontWeight: 600, letterSpacing: "0.05em", color: "var(--fg-3)", textTransform: "uppercase" }}>
-        Config file
-      </div>
-      <div
-        className="mono"
-        style={{ fontSize: "var(--fs-xs)", color: "var(--fg-1)", marginTop: 6, wordBreak: "break-all", lineHeight: 1.4 }}
-      >
-        {path}
-      </div>
-      <button
-        type="button"
-        onClick={copy}
-        aria-label="Copy config path"
-        style={{
-          marginTop: 8,
-          background: "transparent",
-          border: "1px solid var(--border-default)",
-          borderRadius: "var(--r-s)",
-          color: copied ? "var(--ok)" : blocked ? "var(--err)" : "var(--fg-1)",
-          fontSize: "var(--fs-xs)",
-          padding: "3px 8px",
-          cursor: "pointer",
-          display: "inline-flex",
-          alignItems: "center",
-          gap: 4,
-        }}
-      >
-        <span aria-hidden>{copied ? "✓" : blocked ? "⚠" : "⧉"}</span>
-        {copied ? "Copied" : blocked ? "Copy blocked — select manually" : "Copy path"}
-      </button>
-    </div>
-  );
-}
+// ToggleRow, PermColumn, ConfigFileCard extracted to ./_components/
+// as the first slice of the settings-page split.
